Change color temperature range
The Hue bulbs can actually support ranges from 2000K-6500K. This increases the range to 2700K-6500K, matching f.lux, which is probably the most well known and widely used product that has similar goals. 

The ctToRGB function was updated to support color temperatures from 1000K - 40000K. Color is set with Hex if possible because RGB to Hex is a direct conversion. rgbToHSV was also updated. Hue and saturation are set in degrees, and then converted to percent for SmartThings (value is calculated but not used). From my testing with hue bulbs and lightstrips, the colors produced using color temperature and using HSV are almost identical (at least to my eye).