Compare colormode
The Hue API includes a state for colormode: "Note the usage of the
colormode parameter: There are 3 ways of setting the light color: xy,
color temperature (ct) or hue and saturation (hs). A light may contain
different settings for xy, ct and hs, but only the mode indicated by the
colormode parameter will be certain to give the active light color."

Therefore, if the bulb was last adjusted by Hue/Saturation or Hex, and
Circadian Daylight has not calculated a new value for color temperature,
running Circadian Daylight will not update a bulb because the (now
incorrect) ct value will look correct. By looking at colormode we know
if we can trust what is being reported for colorTemperature or color.

This should be written in a way that if a Device Handler which does not
include colormode is being used the functionality will be unchanged from
the behavior before this commit.